Replace solar panel types with battery solar panel behavior.
- Remove obsolete files.
- Solar panels can now only be aligned on the x or y axis, instead of
facing in any horizontal direction.
- Possibly added issues.
