--- YATM/src/main/java/com/gsr/gsr_yatm/block/device/solar/BatterySolarPanelMenu.java
@@ -1,139 +1,0 @@
-package com.gsr.gsr_yatm.block.device.solar;
-
-import com.gsr.gsr_yatm.registry.YATMMenuTypes;
-import com.gsr.gsr_yatm.utilities.capability.SlotUtil;
-
-import net.minecraft.world.entity.player.Inventory;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.inventory.AbstractContainerMenu;
-import net.minecraft.world.inventory.ContainerData;
-import net.minecraft.world.inventory.ContainerLevelAccess;
-import net.minecraft.world.inventory.SimpleContainerData;
-import net.minecraft.world.inventory.Slot;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.block.Block;
-import net.minecraftforge.items.IItemHandler;
-import net.minecraftforge.items.ItemStackHandler;
-import net.minecraftforge.items.SlotItemHandler;
-
-public class BatterySolarPanelMenu extends AbstractContainerMenu
-{
-	public static final int PLAYER_INVENTORY_START = (BatterySolarPanelBlockEntity.INVENTORY_SLOT_COUNT - 1) + 1;
-	public static final int PLAYER_INVENTORY_END = PLAYER_INVENTORY_START + 26;
-	public static final int PLAYER_HOTBAR_START = PLAYER_INVENTORY_END + 1;
-	public static final int PLAYER_HOTBAR_END = PLAYER_HOTBAR_START + 8;
-	
-	private final ContainerLevelAccess m_access;
-	private final Block m_openingBlockType;
-	private final ContainerData m_data;
-	
-	
-	
-	// client side constructor
-	public BatterySolarPanelMenu(int inventoryId, Inventory playerInventory)
-	{
-		this(inventoryId, playerInventory, ContainerLevelAccess.NULL, null, new ItemStackHandler(BatterySolarPanelBlockEntity.INVENTORY_SLOT_COUNT), new SimpleContainerData(BatterySolarPanelBlockEntity.getDataSlotCount()));
-	} // end client constructor
-
-	// server side constructor
-	public BatterySolarPanelMenu(int inventoryId, Inventory playerInventory, ContainerLevelAccess access, Block openingBlockType, IItemHandler objInventory, ContainerData data)
-	{
-		super(YATMMenuTypes.BATTERY_SOLAR_PANEL.get(), inventoryId);
-
-		this.m_access = access;
-		this.m_openingBlockType = openingBlockType;
-		this.m_data = data;
-		this.addSlot(new SlotItemHandler(objInventory, BatterySolarPanelBlockEntity.POWER_SLOT, 79, 51));
-
-		for (int y = 0; y < 3; ++y)
-		{
-			for (int x = 0; x < 9; ++x)
-			{
-				this.addSlot(new Slot(playerInventory, (x + (y * 9)) + 9, 8 + (x * 18), (84) + (y * 18)));
-			}
-		}
-
-		for (int x = 0; x < 9; ++x)
-		{
-			this.addSlot(new Slot(playerInventory, x, 8 + (x * 18), 142));
-		}
-
-		this.addDataSlots(data);
-	} // end server constructor
-
-	
-	
-	@Override
-	public ItemStack quickMoveStack(Player player, int quickMovedSlotIndex)
-	{
-		ItemStack quickMovedStack = ItemStack.EMPTY;
-		Slot quickMovedSlot = this.slots.get(quickMovedSlotIndex);
-		if (quickMovedSlot != null && quickMovedSlot.hasItem())
-		{
-			ItemStack slotsStack = quickMovedSlot.getItem();
-			if (quickMovedSlotIndex >= PLAYER_INVENTORY_START && quickMovedSlotIndex <= PLAYER_HOTBAR_END)
-			{	
-				boolean moved = false;
-				if(SlotUtil.isValidPowerSlotInsert(slotsStack) && this.moveItemStackTo(slotsStack, BatterySolarPanelBlockEntity.POWER_SLOT, BatterySolarPanelBlockEntity.POWER_SLOT + 1, false)) 
-				{					
-					moved = true;
-				}
-				else if((quickMovedSlotIndex >= PLAYER_INVENTORY_START && quickMovedSlotIndex <= PLAYER_INVENTORY_END) && this.moveItemStackTo(slotsStack, PLAYER_HOTBAR_START, PLAYER_HOTBAR_END + 1, false)) 
-				{											
-					moved = true;
-				}
-				else if ((quickMovedSlotIndex >= PLAYER_HOTBAR_START && quickMovedSlotIndex <= PLAYER_HOTBAR_END) && this.moveItemStackTo(slotsStack, PLAYER_INVENTORY_START, PLAYER_INVENTORY_END + 1, false))
-				{
-					moved = true;
-				}
-				if(!moved) 
-				{
-					return ItemStack.EMPTY;
-				}
-			}
-			else if (!this.moveItemStackTo(slotsStack, PLAYER_INVENTORY_START, PLAYER_HOTBAR_END + 1, false))
-			{
-				return ItemStack.EMPTY;
-			}
-			
-			
-			
-			if(slotsStack.isEmpty())
-			{
-				quickMovedSlot.set(ItemStack.EMPTY);
-			}
-			else 
-			{
-				quickMovedSlot.setChanged();
-			}
-			
-			if(slotsStack == quickMovedStack)
-			{
-				return ItemStack.EMPTY;
-			}
-		
-			quickMovedSlot.onTake(player, quickMovedStack);
-		} // end slot usability check
-
-		return quickMovedStack;
-	} // end quickMoveStack()
-
-	@Override
-	public boolean stillValid(Player player)
-	{
-		return AbstractContainerMenu.stillValid(this.m_access, player, this.m_openingBlockType);
-	} // end stillValid
-
-	
-	
-	public int currentStored() 
-	{
-		return this.m_data.get(BatterySolarPanelBlockEntity.getCurrentData().startIndex());
-	} // end currentStored()
-	
-	public int currentCapacity() 
-	{
-		return this.m_data.get(BatterySolarPanelBlockEntity.getCurrentData().endIndex());
-	} // end currentCapacity()
-	
-} // end class
