--- YATM/src/main/java/com/gsr/gsr_yatm/block/device/solar/BatterySolarPanelScreen.java
@@ -1,75 +1,0 @@
-package com.gsr.gsr_yatm.block.device.solar;
-
-import com.gsr.gsr_yatm.YetAnotherTechMod;
-import com.gsr.gsr_yatm.gui.HorizontalCurrentWidget;
-import com.mojang.blaze3d.systems.RenderSystem;
-
-import net.minecraft.client.gui.GuiGraphics;
-import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
-import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.entity.player.Inventory;
-
-public class BatterySolarPanelScreen extends AbstractContainerScreen<BatterySolarPanelMenu>
-{
-	public static final ResourceLocation BACKGROUND = new ResourceLocation(YetAnotherTechMod.MODID, "textures/gui/container/battery_solar_panel.png");
-	
-	private HorizontalCurrentWidget m_currentWidget;
-	
-	
-	
-	public BatterySolarPanelScreen(BatterySolarPanelMenu menu, Inventory inventory, Component title)
-	{
-		super(menu, inventory, title);
-	} // end constructor
-
-	
-	
-	@Override
-	protected void init()
-	{
-		super.init();
-		this.setCurrentWidget();
-	} // end init()
-	
-	@Override
-	public void render(GuiGraphics graphics, int mouseX, int mouseY, float partialTick)
-	{
-		super.renderBackground(graphics, mouseX, mouseY, partialTick);
-		this.renderBg(graphics, partialTick, mouseX, mouseY);
-		this.updateCurrentWidget();
-		super.render(graphics, mouseX, mouseY, partialTick);
-		this.renderTooltip(graphics, mouseX, mouseY);
-	} // end render()
-
-	@Override
-	protected void renderBg(GuiGraphics graphics, float partialTick, int mouseX, int mouseY)
-	{
-		RenderSystem.setShaderTexture(0, BACKGROUND);
-		graphics.blit(BACKGROUND, this.leftPos, this.topPos, 0, 0, this.imageWidth, this.imageHeight);
-	} // end renderBg()
-
-	
-	
-	public void updateCurrentWidget() 
-	{
-		if(this.m_currentWidget == null || this.menu.currentCapacity() != this.m_currentWidget.getCapacity()) 
-		{
-			this.setCurrentWidget();
-		}
-
-		this.m_currentWidget.setStored(this.menu.currentStored());
-	} // end updateResultTankWidget()
-	
-	public void setCurrentWidget() 
-	{
-		if(this.m_currentWidget != null) 
-		{			
-			this.removeWidget(this.m_currentWidget);
-		}
-		
-		this.m_currentWidget = new HorizontalCurrentWidget(this.leftPos + 43, this.topPos + 21, this.getMenu().currentCapacity());
-		this.addRenderableWidget(this.m_currentWidget);
-	} // end setWidget()
-	
-} // end class
