+++ YATM/src/main/java/com/gsr/gsr_yatm/block/device/solar/SolarPanelBlockEntity.java
@@ -2,35 +2,25 @@ package com.gsr.gsr_yatm.block.device.solar;
 
 import com.google.common.collect.ImmutableList;
 import com.gsr.gsr_yatm.registry.YATMBlockEntityTypes;
+
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
-import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 
 public class SolarPanelBlockEntity extends AbstractSolarPanelBlockEntity
 {
-	private static final ImmutableList<Direction> POWERABLE_FACES = ImmutableList.of(Direction.DOWN);
+	private static final ImmutableList<Direction> POWERABLE_FACES = ImmutableList.of(Direction.DOWN, Direction.NORTH, Direction.SOUTH, Direction.EAST, Direction.WEST);
 	
 	
 	
-	public SolarPanelBlockEntity(BlockPos blockPos, BlockState blockState)
+	public SolarPanelBlockEntity(BlockPos position, BlockState state)
 	{
-		super(YATMBlockEntityTypes.SOLAR_PANEL.get(), blockPos, blockState, POWERABLE_FACES);
+		super(YATMBlockEntityTypes.SOLAR_PANEL.get(), position, state, POWERABLE_FACES);
 	} // end constructor
 	
-	public SolarPanelBlockEntity(BlockPos blockPos, BlockState blockState, int currentCapacity, int maxSafeCurrentTransfer, int maxCurrentTransfer, SolarPanelSettings settings)
+	public SolarPanelBlockEntity(BlockPos position, BlockState state, int currentCapacity, int maxSafeCurrentTransfer, int maxCurrentTransfer, SolarPanelSettings settings)
 	{
-		super(YATMBlockEntityTypes.SOLAR_PANEL.get(), blockPos, blockState, SolarPanelBlockEntity.POWERABLE_FACES, currentCapacity, maxSafeCurrentTransfer, maxCurrentTransfer, settings);
+		super(YATMBlockEntityTypes.SOLAR_PANEL.get(), position, state, POWERABLE_FACES, currentCapacity, maxSafeCurrentTransfer, maxCurrentTransfer, settings);
 	} // end constructor
 
-	
-	
-	@Override
-	public void serverTick(Level level, BlockPos position, BlockState state)
-	{
-		// hopefully this successfully clears it
-		this.m_internalCurrentStorer.extractCurrent(Integer.MAX_VALUE, false);
-		super.serverTick(level, position, state);
-	} // end serverTick()
-	
 } // end class
