+++ YATM/src/main/java/com/gsr/gsr_yatm/block/device/solar/SolarPanelScreen.java
@@ -1,6 +1,7 @@
 package com.gsr.gsr_yatm.block.device.solar;
 
 import com.gsr.gsr_yatm.YetAnotherTechMod;
+import com.gsr.gsr_yatm.gui.HorizontalCurrentWidget;
 import com.mojang.blaze3d.systems.RenderSystem;
 
 import net.minecraft.client.gui.GuiGraphics;
@@ -13,6 +14,8 @@ public class SolarPanelScreen extends AbstractContainerScreen<SolarPanelMenu>
 {
 	public static final ResourceLocation BACKGROUND = new ResourceLocation(YetAnotherTechMod.MODID, "textures/gui/container/solar_panel.png");
 	
+	private HorizontalCurrentWidget m_currentWidget;
+	
 	
 	
 	public SolarPanelScreen(SolarPanelMenu menu, Inventory inventory, Component title)
@@ -23,10 +26,18 @@ public class SolarPanelScreen extends AbstractContainerScreen<SolarPanelMenu>
 	
 	
 	@Override
+	protected void init()
+	{
+		super.init();
+		this.setCurrentWidget();
+	} // end init()
+	
+	@Override
 	public void render(GuiGraphics graphics, int mouseX, int mouseY, float partialTick)
 	{
 		super.renderBackground(graphics, mouseX, mouseY, partialTick);
 		this.renderBg(graphics, partialTick, mouseX, mouseY);
+		this.updateCurrentWidget();
 		super.render(graphics, mouseX, mouseY, partialTick);
 		this.renderTooltip(graphics, mouseX, mouseY);
 	} // end render()
@@ -34,8 +45,31 @@ public class SolarPanelScreen extends AbstractContainerScreen<SolarPanelMenu>
 	@Override
 	protected void renderBg(GuiGraphics graphics, float partialTick, int mouseX, int mouseY)
 	{
-		RenderSystem.setShaderTexture(0, SolarPanelScreen.BACKGROUND);
-		graphics.blit(SolarPanelScreen.BACKGROUND, this.leftPos, this.topPos, 0, 0, this.imageWidth, this.imageHeight);
+		RenderSystem.setShaderTexture(0, BACKGROUND);
+		graphics.blit(BACKGROUND, this.leftPos, this.topPos, 0, 0, this.imageWidth, this.imageHeight);
 	} // end renderBg()
+
+	
+	
+	public void updateCurrentWidget() 
+	{
+		if(this.m_currentWidget == null || this.menu.currentCapacity() != this.m_currentWidget.getCapacity()) 
+		{
+			this.setCurrentWidget();
+		}
+
+		this.m_currentWidget.setStored(this.menu.currentStored());
+	} // end updateResultTankWidget()
+	
+	public void setCurrentWidget() 
+	{
+		if(this.m_currentWidget != null) 
+		{			
+			this.removeWidget(this.m_currentWidget);
+		}
+		
+		this.m_currentWidget = new HorizontalCurrentWidget(this.leftPos + 43, this.topPos + 21, this.getMenu().currentCapacity());
+		this.addRenderableWidget(this.m_currentWidget);
+	} // end setWidget()
 	
 } // end class
