+++ YATM/src/main/java/com/gsr/gsr_yatm/registry/YATMMenuTypes.java
@@ -13,7 +13,6 @@ import com.gsr.gsr_yatm.block.device.crafting.injector.InjectorMenu;
 import com.gsr.gsr_yatm.block.device.crafting.still.StillMenu;
 import com.gsr.gsr_yatm.block.device.creative.current_source.CreativeCurrentSourceMenu;
 import com.gsr.gsr_yatm.block.device.current_storer.base.CurrentStorerMenu;
-import com.gsr.gsr_yatm.block.device.solar.BatterySolarPanelMenu;
 import com.gsr.gsr_yatm.block.device.solar.SolarPanelMenu;
 import com.gsr.gsr_yatm.item.tool.PoweredToolMenu;
 
@@ -43,7 +42,6 @@ public class YATMMenuTypes
       
 	public static final RegistryObject<MenuType<CurrentStorerMenu>> CURRENT_STORER = MENU_TYPES.register("current_storer", () -> new MenuType<>(CurrentStorerMenu::new, FeatureFlagSet.of())); 
 	
-	public static final RegistryObject<MenuType<BatterySolarPanelMenu>> BATTERY_SOLAR_PANEL = MENU_TYPES.register("battery_solar_panel", () -> new MenuType<>(BatterySolarPanelMenu::new, FeatureFlagSet.of()));
 	public static final RegistryObject<MenuType<SolarPanelMenu>> SOLAR_PANEL = MENU_TYPES.register("solar_panel", () -> new MenuType<>(SolarPanelMenu::new, FeatureFlagSet.of()));
 
 	public static final RegistryObject<MenuType<CreativeCurrentSourceMenu>> CREATIVE_CURRENT_SOURCE = MENU_TYPES.register("creative_current_source", () -> new MenuType<>(CreativeCurrentSourceMenu::new, FeatureFlagSet.of()));
